Create dataframe with values of each variable on each site

Create dataframe with values of each variable on each site

data_JSDM(EPSG = 3163, latlon_site, area_site, path_tiff, log_area_site = TRUE)

Arguments

EPSG

int. target EPSG, default 3163

latlon_site

float matrix with 2 columns. Columns are longitude and latitude in this order.

area_site

float vector same dimension then number of row `latlon`. area of each site in m²

path_tiff

character. path to Tiff file with explanatories variables

log_area_site

boolean. Replace area of each site by log area of each site. Anyway, this variable is scale.

Value

matrix with latitude, longitude and all variables. If `latlon_site` has row names, output has same row name.